Break out filesystem watcher into separate file
EventServiceManager is very large.  Break out two of the functions, and
the global variables into a separate compile unit.

Code is copied as-is, with no improvements made in this patch.

Tested: At end of series.
