+++ meson.build
@@ -334,6 +334,7 @@ srcfiles_bmcweb = files(
 
     'redfish-core/src/error_messages.cpp',
     'redfish-core/src/event_log.cpp',
+    'redfish-core/src/filesystem_log_watcher.cpp',
     'redfish-core/src/filter_expr_executor.cpp',
     'redfish-core/src/filter_expr_printer.cpp',
     'redfish-core/src/redfish.cpp',
