+++ redfish-core/include/filesystem_log_watcher.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <boost/asio/io_context.hpp>
+
+namespace redfish
+{
+int startEventLogMonitor(boost::asio::io_context& ioc);
+void stopEventLogMonitor();
+} // namespace redfish
