+++ src/webserver_run.cpp
@@ -6,6 +6,7 @@
 #include "dbus_monitor.hpp"
 #include "dbus_singleton.hpp"
 #include "event_service_manager.hpp"
+#include "filesystem_log_watcher.hpp"
 #include "google/google_service_root.hpp"
 #include "hostname_monitor.hpp"
 #include "ibm/management_console_rest.hpp"
@@ -125,7 +126,7 @@ int run()
 
     if constexpr (!BMCWEB_REDFISH_DBUS_LOG)
     {
-        int rc = redfish::EventServiceManager::startEventLogMonitor(*io);
+        int rc = redfish::startEventLogMonitor(*io);
         if (rc != 0)
         {
             BMCWEB_LOG_ERROR("Redfish event handler setup failed...");
@@ -147,7 +148,7 @@ int run()
     crow::connections::systemBus = nullptr;
 
     // TODO(ed) Make event log monitor an RAII object instead of global vars
-    redfish::EventServiceManager::stopEventLogMonitor();
+    redfish::stopEventLogMonitor();
 
     return 0;
 }
